src/libstatgrab/cpu_stats.c: avoid vmmeter compile warnings
Add explicitely type conversion to (unsigned long long) from counter_u64_t
to avoid compile warnings:
```
cpu_stats.c:347:35: warning: incompatible pointer to integer conversion assigning to 'unsigned long long' from 'counter_u64_t' (aka 'unsigned long *') [-Wint-conversion]
                cpu_stats_buf->context_switches = vmmeter.v_swtch;
                                                ^ ~~~~~~~~~~~~~~~
cpu_stats.c:348:27: warning: incompatible pointer to integer conversion assigning to 'unsigned long long' from 'counter_u64_t' (aka 'unsigned long *') [-Wint-conversion]
                cpu_stats_buf->syscalls = vmmeter.v_syscall;
                                        ^ ~~~~~~~~~~~~~~~~~
cpu_stats.c:349:27: warning: incompatible pointer to integer conversion assigning to 'unsigned long long' from 'counter_u64_t' (aka 'unsigned long *') [-Wint-conversion]
                cpu_stats_buf->syscalls += vmmeter.v_trap;
                                        ^  ~~~~~~~~~~~~~~
cpu_stats.c:351:29: warning: incompatible pointer to integer conversion assigning to 'unsigned long long' from 'counter_u64_t' (aka 'unsigned long *') [-Wint-conversion]
                cpu_stats_buf->interrupts = vmmeter.v_intr;
                                          ^ ~~~~~~~~~~~~~~
cpu_stats.c:352:34: warning: incompatible pointer to integer conversion assigning to 'unsigned long long' from 'counter_u64_t' (aka 'unsigned long *') [-Wint-conversion]
                cpu_stats_buf->soft_interrupts = vmmeter.v_soft;
                                               ^ ~~~~~~~~~~~~~~
5 warnings generated.
```

+++ src/libstatgrab/cpu_stats.c
@@ -344,12 +344,12 @@ sg_get_cpu_stats_int(sg_cpu_stats *cpu_stats_buf) {
 	size = sizeof(vmmeter);
 	memset(&vmmeter, 0, sizeof(vmmeter));
 	if( (sysctl( mib, 2, &vmmeter, &size, NULL, 0 ) == 0) && (size == sizeof(vmmeter)) ) {
-		cpu_stats_buf->context_switches = vmmeter.v_swtch;
-		cpu_stats_buf->syscalls = vmmeter.v_syscall;
-		cpu_stats_buf->syscalls += vmmeter.v_trap;
+		cpu_stats_buf->context_switches = (unsigned long long)vmmeter.v_swtch;
+		cpu_stats_buf->syscalls = (unsigned long long)vmmeter.v_syscall;
+		cpu_stats_buf->syscalls += (unsigned long long)vmmeter.v_trap;
 		cpu_stats_buf->voluntary_context_switches = cpu_stats_buf->involuntary_context_switches = 0;
-		cpu_stats_buf->interrupts = vmmeter.v_intr;
-		cpu_stats_buf->soft_interrupts = vmmeter.v_soft;
+		cpu_stats_buf->interrupts = (unsigned long long)vmmeter.v_intr;
+		cpu_stats_buf->soft_interrupts = (unsigned long long)vmmeter.v_soft;
 	}
 	else {
 		cpu_stats_buf->context_switches = 0;
